backend/requirements.txt
fastapi==0.110.1
uvicorn==0.27.1
sqlalchemy==2.0.25
asyncpg==0.29.0
httpx==0.26.0
python-dotenv==1.0.1
PyYAML==6.0.1
prometheus-client==0.19.0
pydantic==2.6.4
pydantic-settings==2.1.0
loguru==0.7.2
alembic==1.13.1
cachetools==5.3.3
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==4.1.0
uvloop==0.19.0
